Add framework for barclamps to save config for use in other barclamps
The goal is to slowly get rid of search in cookbooks and instead have a
backend (right now, chef data bags) for config bits where these config
bits can be read directly.

Rationale for this is that depending on search doesn't really work well
in practice, as nodes might move to a state where some roles are not
active anymore (for instance, when a node enters "problem", it has
no role except the basic ones anymore). This means that configuration
that we generate based on the result of search depends on the state of
the nodes, which is wrong.

Later on, we can also replace the backend to something else than data
bags and remove one more reason for depending on chef-server.